+++ src/app/app.component.ts
@@ -3,303 +3,24 @@ import { CommonModule } from '@angular/common';
 import { RouterOutlet } from '@angular/router';
 
 import { MenuComponent } from './menu/menu.component';
+import { HeroComponent } from './hero/hero.component';
+import { ResumeComponent } from './resume/resume.component';
+import { TransitionComponent } from './transition/transition.component';
+import { HighschoolComponent } from './highschool/highschool.component';
+import { IutComponent } from './iut/iut.component';
+import { CountactComponent } from './countact/countact.component';
+import { ScuComponent } from './scu/scu.component';
+import { ProjectsComponent } from './projects/projects.component';
+import { WorkwithmeComponent } from './workwithme/workwithme.component';
 import { FooterComponent } from './footer/footer.component';
-import { Fancybox } from "@fancyapps/ui";
-import "@fancyapps/ui/dist/fancybox/fancybox.css";
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [CommonModule, RouterOutlet, MenuComponent, FooterComponent],
+  imports: [CommonModule, RouterOutlet, MenuComponent, FooterComponent, HeroComponent, ResumeComponent, TransitionComponent, HighschoolComponent, IutComponent, CountactComponent, ScuComponent, ProjectsComponent, WorkwithmeComponent],
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
-  title = 'portfolio';
-  container : HTMLElement | null;
 
-  active : boolean;
-  currentX : number | null;
-  currentY : number | null;
-  initialX : number | null;
-  initialY : number | null;
-  itemClick : HTMLElement | null;
-  xOffset : number;
-  yOffset : number;
-  typewriterWords: string[] | null;
-
-  constructor() {
-    this.active = false;
-    this.currentX = null;
-    this.currentY = null;
-    this.initialX = null;
-    this.initialY = null;
-    this.itemClick = null;
-    this.xOffset = 0;
-    this.yOffset = 0;
-
-    this.typewriterWords = ["WEBSITE", "SERVER", "MOBILE APPLICATION", "DESKTOP APPLICATION", "GAME"];
-
-    this.container = document.getElementById("HSContainer");
-  }
-
-  // TypeWriter
-  typeWriter(word: string, i: number, text: string) {
-    if (i < word.length) {
-      text += word.charAt(i);
-      document.getElementById("Typewriter")!.innerHTML = text;
-      i++;
-      setTimeout(() => this.typeWriter(word, i, text), 100);
-    }else{
-      setTimeout(() => this.typeWriterBack(word, i, text), 1000);
-    }
-  }
-
-  typeWriterBack(word: string, i: number, text: string) {
-    if (i > 0) {
-      text = text.slice(0, -1);
-      document.getElementById("Typewriter")!.innerHTML = text;
-      i--;
-      setTimeout(() => this.typeWriterBack(word, i, text), 50);
-    }else{
-      // Change the word
-      let index = this.typewriterWords!.indexOf(word);
-      index = (index + 1) % this.typewriterWords!.length;
-      word = this.typewriterWords![index];
-      setTimeout(() => this.typeWriter(word, i, text), 1000);
-    }
-  }
-
-  // Loop through the words
-  typeWriterLoop() {
-    let i = 0;
-    let text = "";
-    let word = this.typewriterWords![i];
-    this.typeWriter(word, 0, text);
-  }
-
-  ngOnInit(): void {
-    let resumeSection = document.getElementById("Resume");
-    let resumeImage = document.getElementById("ResumeImage");
-    let resumerClickable = document.getElementById("ResumeClickable");
-    let resumelightbox = document.getElementById("ResumeLightbox");
-    resumeSection?.addEventListener("click", function() {
-      resumeImage?.classList.toggle("expand");
-      resumeSection?.classList.toggle("expand");
-      resumerClickable?.classList.toggle("expand");
-    });
-    resumerClickable?.addEventListener("click", function() {
-      resumerClickable?.classList.contains("expand") ? null : null;
-    });
-    resumelightbox?.addEventListener("click", function() {
-      Fancybox.bind("[data-fancybox]", {
-        // Your custom options
-      });
-    });
-
-    dragElement(document.getElementById("HSISN")!);
-    dragElement(document.getElementById("HSKart")!);
-    dragElement(document.getElementById("HSMinecraft")!);
-
-    // Get the IUT Containers
-    var IUTContainer = document.getElementsByClassName("IUTSlider") as HTMLCollectionOf<HTMLElement>;
-    // Get the IUT Text
-    var IUTText = document.getElementsByClassName("IUTText") as HTMLCollectionOf<HTMLElement>;
-    // Add an event listener to the IUT Containers when the mouse is over
-    // Change the opacity of the text to 1
-    for (let i = 0; i < IUTContainer.length; i++) {
-      IUTContainer[i].addEventListener("mouseover", function() {
-        IUTText[i].style.opacity = "1";
-        // Set font size to 1.5rem
-        IUTText[i].style.fontSize = "1em";
-      });
-      IUTContainer[i].addEventListener("mouseout", function() {
-        IUTText[i].style.opacity = "0";
-        // Set font size to 0rem
-        IUTText[i].style.fontSize = "0em";
-      });
-    }
-
-    // Use the typewriter
-    this.typeWriterLoop();
-
-    // Get both parts of the countact logo
-    var countactLogoLeft = document.getElementById("CountactLeftSlider") as HTMLElement;
-    var countactLogoRight = document.getElementById("CountactRightSlider") as HTMLElement;
-
-    // Get both texts of the countact logo
-    var countactTextLeft = document.getElementById("CountactLeftText") as HTMLElement;
-    var countactTextRight = document.getElementById("CountactRightText") as HTMLElement;
-  
-    // Add event listeners to clicks
-    countactLogoLeft.addEventListener("click", function() {
-      if (!countactLogoLeft.classList.contains("hidden")) {
-        countactLogoLeft.classList.toggle("shown");
-        countactTextLeft.hidden = !countactTextLeft.hidden;
-        countactLogoRight.classList.toggle("hidden");
-      }
-    });
-    countactLogoRight.addEventListener("click", function() {
-      if (!countactLogoRight.classList.contains("hidden")) {
-        countactLogoLeft.classList.toggle("hidden");
-        countactLogoRight.classList.toggle("shown");
-        countactTextRight.hidden = !countactTextRight.hidden;
-      }
-    });
-
-    // Projects elements
-    // Get the project elements
-    var projects_slider_items = document.getElementsByClassName("ProjectsSliderItem") as HTMLCollectionOf<HTMLElement>;
-
-    // Add a transform to each element, the first 0% on x, then 20% per element
-    for (let i = 0; i < projects_slider_items.length; i++) {
-      projects_slider_items[i].style.transform = "translateX(" + (i * 20) + "%)";
-    }
-
-    // Add an event listener to the projects slider item, using the dragProjectElement function
-    for (let i = 0; i < projects_slider_items.length; i++) {
-      projects_slider_items[i].addEventListener("mousedown", function() {
-        dragProjectElement(projects_slider_items[i]);
-      });
-    }
-
-    // Parallax effect
-    // Get the three images
-    var TI1 = document.getElementById("TI1") as HTMLElement;
-    var TI2 = document.getElementById("TI2") as HTMLElement;
-    var TI3 = document.getElementById("TI3") as HTMLElement;
-    // Event listener for the scroll event
-    window.addEventListener("scroll", function() {
-      // Get the scroll position
-      var scroll = window.scrollY;
-      // Set the transform to the images
-      TI1.style.transform = "translateX(" + (scroll<1445 ? (scroll / 3 - 450) : 31.6) + "%) translateY(-60%)";
-      // Rise opacity of the image from 0% when scroll is 500 to 100% whe nthe scroll gets to 1445
-      TI2.style.transform = "translateY(" + (scroll>1120 ? (scroll<1445 ? (scroll / 2 - 722) : 0) : -161.5) + "%) translateX(-" + (scroll / 700) + "%)";
-      TI3.style.transform = "translateY(" + (scroll<1445 ? (-scroll / 4 + 410) : 50) + "%) translateX(-40%)"; 
-      TI1.style.opacity = scroll/6 - 140 + "%";
-      TI2.style.opacity = scroll/6 - 140 + "%";
-      TI3.style.opacity = scroll/6 - 140 + "%";
-    });
-  }
-}
-
-function dragProjectElement(elmnt: HTMLElement) {
-  console.log("Drag project element");
-  console.log(elmnt);
-  var pos1 = 0, pos3 = 0;
-  elmnt.onmousedown = dragMouseDown;
-
-  function dragMouseDown(e: any) {
-    e = e || window.event;
-    e.preventDefault();
-    // get the mouse cursor position at startup:
-    pos3 = e.clientX;
-    elmnt.style.transition = "none";
-    // document.onmouseup = closeDragElement;
-    // // call a function whenever the cursor moves:
-    // document.onmousemove = elementDrag;
-  }
-}
-
-function dragElement(elmnt: HTMLElement) {
-  var pos1 = 0, pos3 = 0;
-  elmnt.onmousedown = dragMouseDown;
-
-  function dragMouseDown(e: any) {
-    e = e || window.event;
-    e.preventDefault();
-    // get the mouse cursor position at startup:
-    pos3 = e.clientX;
-    elmnt.style.transition = "none";
-    document.onmouseup = closeDragElement;
-    // call a function whenever the cursor moves:
-    document.onmousemove = elementDrag;
-  }
-
-  function elementDrag(e: any) {
-    e = e || window.event;
-    e.preventDefault();
-    // calculate the new cursor position:
-    pos1 = pos3 - e.clientX;
-    pos3 = e.clientX;
-
-    // Get the 
-
-    let elmntNewPos = elmnt.offsetLeft - pos1;
-
-    // Get the container element
-    var container = elmnt.parentElement;
-    // Get the offset position of the container
-    var containerRect = container!.getBoundingClientRect();
-    // Get the width of the container
-    var containerWidth = containerRect.width;
-    // Get the width of the element
-    // If the element has the text shown, substract the width of the text
-    if (elmnt.classList.contains("shown")) {
-      var text = elmnt.getElementsByClassName("HSText")[0] as HTMLElement;
-      var elmntWidth = elmnt.getBoundingClientRect().width - text.getBoundingClientRect().width;
-    }else{
-      var elmntWidth = elmnt.getBoundingClientRect().width;
-    }
-
-    if(elmntNewPos < containerRect.left || elmntNewPos > containerRect.left + containerWidth - elmntWidth) {
-      closeDragElement();
-    }else{
-      elmnt.style.left = (elmntNewPos) + "px";
-    }
-  }
-
-  function closeDragElement() {
-    // stop moving when mouse button is released:
-    document.onmouseup = null;
-    document.onmousemove = null;
-
-    // Get the container element
-    var container = elmnt.parentElement;
-    // Get the offset position of the container
-    var containerRect = container!.getBoundingClientRect();
-    // Get the width of the container
-    var containerWidth = containerRect.width;
-    // Get the width of the element
-    if (elmnt.classList.contains("shown")) {
-      var text = elmnt.getElementsByClassName("HSText")[0] as HTMLElement;
-      var elmntWidth = elmnt.getBoundingClientRect().width - text.getBoundingClientRect().width;
-    }else{
-      var elmntWidth = elmnt.getBoundingClientRect().width;
-    }
-
-    // Passing point
-    var passingPoint = containerRect.left + (containerWidth / 2) - (elmntWidth / 2);
-
-    // Set the element's new position:
-    if (elmnt.offsetLeft > passingPoint) {
-      elmnt.style.transition = "all .2s cubic-bezier(0.04, 0.46, 0.36, 0.99)";
-      elmnt.style.left = containerRect.left + containerWidth - elmntWidth + "px";
-
-      // Toggle the classlist shown to the element
-      elmnt.classList.remove("shown");
-    } else {
-      elmnt.style.transition = "all .2s cubic-bezier(0.04, 0.46, 0.36, 0.99)";
-      elmnt.style.left = containerRect.left + "px";
-
-      // Toggle the classlist shown to the element
-      elmnt.classList.add("shown");
-    }
-
-    // After the new position is done, show the text if the element is shown
-    if (elmnt.classList.contains("shown")) {
-      
-      // Get the text element
-      var text = elmnt.getElementsByClassName("HSText")[0] as HTMLElement;
-      // Change the style of the text element to remove the hidden class
-      text.hidden = false;
-
-    }else{
-      // Get the text element
-      var text = elmnt.getElementsByClassName("HSText")[0] as HTMLElement;
-      // Change the style of the text element to remove the hidden class
-      text.hidden = true;
-    }
-  }
 }
