+++ src/app/countact/countact.component.css
@@ -1,0 +1,98 @@
+#Countact{
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+    margin: auto;
+    overflow-y: clip;
+    overflow-x: clip;
+    align-items: center;
+
+    background-image: url("../../assets/images/Low-Opacity/countact-low.png");
+    background-size: cover;
+    background-position-x: center;
+    background-position-y:bottom;
+    background-attachment: fixed;
+}
+
+#Countact .subTitle{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+#Countact .subTitle h2{
+    margin: 0;
+}
+
+#CountactTopWhiteBlur{
+    width: 150%;
+    height: 10%;
+    background-color: var(--quinary-color);
+    filter: blur(25px);
+    background-size: cover;
+    transform: translateX(-10%) translateY(-50%);
+}
+
+#CountactImage{
+    object-fit: cover;
+    opacity: 0.2;
+    position: absolute;
+    width: 100%;
+    height: 100vh;
+    z-index: -1;
+}
+
+#CountactBottomWhiteBlur{
+    width: 150%;
+    height: 10%;
+    background-color: var(--quinary-color);
+    filter: blur(25px);
+    background-size: cover;
+    transform: translateX(-10%) translateY(50%);
+}
+
+#CountactBody{
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+
+#CountactSlider{
+    display: flex;
+    max-width: 100%;
+    align-items: center;
+    align-self: center;
+    height: 60%;
+}
+
+#CountactSlider > div{
+    width: 100%;
+    height: 70%;
+    display: flex;
+    align-items: center;
+    transition: all 0.5s ease-in-out;
+}
+
+#CountactSlider > div > img{
+    max-width: 100%;
+    height: 100%;
+    cursor: pointer;
+}
+
+#CountactLeftSlider{
+    justify-content: flex-end;
+}
+
+#CountactLeftText, #CountactRightText{
+    margin: 0 5%;
+    max-width: 50%;
+    transition: all 0.5s ease-in-out;
+}
+
+.hidden{
+    opacity: 0%;
+}
