+++ src/app/countact/countact.component.html
@@ -1,0 +1,43 @@
+<div id="Countact">
+    <!-- <div id="CountactTopWhiteBlur"></div> -->
+    
+    <!-- <img id="CountactImage" src="assets/images/countact-low.png" alt="Countact Image"> -->
+  
+    <div id="CountactBody">
+      <div class="subTitle">
+        <h2>Internships Countact</h2>
+        <h3>Year 2022-2023</h3>
+      </div>
+      <!-- Add the sliders -->
+      <div id="CountactSlider">
+        <div id="CountactLeftSlider">
+          <img src="../assets/images/countact_logo_left.png">
+          <div id="CountactLeftText" hidden>
+            <div>
+              <h3>Countact</h3>
+              <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum lacinia efficitur condimentum. Duis aliquet rhoncus diam, eu tempor lacus ullamcorper at. Duis at luctus dui. Praesent ultrices blandit ante, ut dapibus sem aliquet a. Nam pharetra turpis velit, et bibendum sapien pellentesque et. Phasellus odio diam, consectetur quis tincidunt vel, porta nec arcu.</p>
+            </div>
+            <div>
+              <h3>Countact</h3>
+              <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum lacinia efficitur condimentum. Duis aliquet rhoncus diam, eu tempor lacus ullamcorper at. Duis at luctus dui. Praesent ultrices blandit ante, ut dapibus sem aliquet a. Nam pharetra turpis velit, et bibendum sapien pellentesque et. Phasellus odio diam, consectetur quis tincidunt vel, porta nec arcu.</p>
+            </div>
+          </div>
+        </div>
+        <div id="CountactRightSlider">
+          <div id="CountactRightText" hidden>
+            <div>
+              <h3>Countact</h3>
+              <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum lacinia efficitur condimentum. Duis aliquet rhoncus diam, eu tempor lacus ullamcorper at. Duis at luctus dui. Praesent ultrices blandit ante, ut dapibus sem aliquet a. Nam pharetra turpis velit, et bibendum sapien pellentesque et. Phasellus odio diam, consectetur quis tincidunt vel, porta nec arcu.</p>
+            </div>
+            <div>
+              <h3>Countact</h3>
+              <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum lacinia efficitur condimentum. Duis aliquet rhoncus diam, eu tempor lacus ullamcorper at. Duis at luctus dui. Praesent ultrices blandit ante, ut dapibus sem aliquet a. Nam pharetra turpis velit, et bibendum sapien pellentesque et. Phasellus odio diam, consectetur quis tincidunt vel, porta nec arcu.</p>
+            </div>
+          </div>
+          <img src="../assets/images/countact_logo_right.png">
+        </div>
+      </div>
+    </div>
+  
+    <!-- <div id="CountactBottomWhiteBlur"></div> -->
+  </div>
