+++ src/app/hero/hero.component.css
@@ -1,0 +1,43 @@
+#Hero{
+    width: 100%;
+    height: 100vh;
+    max-width:100%;
+    overflow-y: clip;
+    overflow-x: clip;
+
+    background-image: url("../../assets/images/Low-Opacity/Hero.png");
+    background-size: cover;
+    background-position: center;
+    background-attachment: fixed;
+}
+
+#HeroImage{
+    width: 100%;
+    height: 100vh;
+    object-fit: cover;
+    opacity: 0.8;
+    position: absolute;
+}
+
+#HeroText{
+    position: absolute;
+    font-size: 2em;
+    max-width: 75%;
+    left: 5%;
+    top: 20%;
+    text-align: left;
+    font-family: 'Roboto', sans-serif;
+}
+
+#Typewriter{
+    color: var(--quaternary-color);
+}
+
+#HeroWhiteBlur{
+    width: 150%;
+    height: 50%;
+    transform: translateX(-10%) translateY(185%);
+    background-color: var(--quinary-color);
+    filter: blur(30px);
+    background-size: cover;
+}
