+++ src/app/hero/hero.component.html
@@ -1,0 +1,8 @@
+<div id="Hero">
+    <!-- <img id="HeroImage" src="assets/images/Hero.jpg" alt="Hero Image"> -->
+    <div id="HeroText">
+      <h1>HI, I'M <span id="HeroName">SYMEON</span></h1>
+      <h3>READY TO WORK ON YOUR <span id="Typewriter"></span></h3>
+    </div>
+    <!-- <div id="HeroWhiteBlur"></div> -->
+  </div>
