+++ src/app/highschool/highschool.component.html
@@ -1,0 +1,56 @@
+<div id="HighSchool">
+    <!-- <div id="HighSchoolTopWhiteBlur"></div> -->
+    
+    <!-- <img id="HighSchoolImage" src="assets/images/HighSchool-low.png" alt="High School Image"> -->
+      
+    <div id="HighSchoolText">
+      <div class="subTitle">
+        <h2>Lycée Pablo Neruda</h2>
+        <h3>Year 2017-2020 <br> Baccalauréat scientifique - Mention très bien</h3>
+      </div>
+  
+      <div id="HighSchoolSliders">
+        <div class="HSContainer">
+          <div id="HSISN">
+            <div>
+              <img src="../assets/images/logo_ISN.png">
+              <h3 class="HSTitle">ISN</h3>
+            </div>
+            <div class="HSText" hidden>
+              <p>ISN text</p>
+              <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur fermentum odio at nibh pretium mollis. Duis mollis sapien ut nulla suscipit, in posuere purus accumsan. Sed vitae justo et nulla elementum porttitor. Praesent sit amet augue sapien.</p>
+            </div>
+          </div>
+        </div>
+        
+        <div class="HSContainer">
+          <div id="HSKart">
+            <div>
+              <img src="../assets/images/logo_kart.png">
+              <h3 class="HSTitle">Kart</h3>
+            </div>
+            <div class="HSText" hidden>
+              <p>Kart text</p>
+              <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur fermentum odio at nibh pretium mollis. Duis mollis sapien ut nulla suscipit, in posuere purus accumsan. Sed vitae justo et nulla elementum porttitor. Praesent sit amet augue sapien.</p>
+            </div>
+          </div>
+        </div>
+  
+        <div class="HSContainer">
+          <div id="HSMinecraft">
+            <div>
+              <img src="../assets/images/logo_minecraft.png">
+              <h3 class="HSTitle">Minecraft</h3>
+            </div>
+            <div class="HSText" hidden>
+              <p>Minecraft text</p>
+              <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur fermentum odio at nibh pretium mollis. Duis mollis sapien ut nulla suscipit, in posuere purus accumsan. Sed vitae justo et nulla elementum porttitor. Praesent sit amet augue sapien.</p>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+    
+  
+    <!-- <div id="HighSchoolBottomWhiteBlur"></div> -->
+  </div>
