+++ src/app/highschool/highschool.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { HighschoolComponent } from './highschool.component';
+
+describe('HighschoolComponent', () => {
+  let component: HighschoolComponent;
+  let fixture: ComponentFixture<HighschoolComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [HighschoolComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(HighschoolComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
