+++ src/app/iut/iut.component.css
@@ -1,0 +1,81 @@
+#IUT{
+    width: 95%;
+    height: 70vh;
+    display: flex;
+    margin: auto;
+    flex-direction: row;
+    overflow-y: clip;
+    overflow-x: clip;
+    justify-content: space-between;
+    align-items: center;
+}
+
+#IUT .subTitle{
+    display: flex;
+    flex-direction: column;
+    justify-content: start;
+    align-content: center;
+
+    height: 50%;
+    margin-top: 10%;
+    margin-right: 5%;
+}
+
+.IUTSliderContainer{
+    display: flex;
+    flex-direction: row;
+    justify-content: end;
+    align-content: center;
+    height: 45vh;
+    width: 100%;
+    max-width: 1000px;
+}
+
+.IUTSlider{
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    height: 100%;
+    width: 25%;
+    transition: all 0.5s ease-in-out;
+    background-size: cover;
+}
+
+.IUTSlider.Classproject{
+    background-image: url("../../assets/images/class_project.png");
+}
+
+.IUTSlider.Graduationproject{
+    background-image: url("../../assets/images/graduation_project.png");
+}
+
+.IUTSlider.Bdeproject{
+    background-image: url("../../assets/images/bde_project.png");
+}
+
+.IUTSlider:hover{
+    width: 50%;
+    background-color: rgba(0,0,0,0.1);
+}
+
+.IUTSlider img{
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    object-position: top;
+}
+
+.IUTText{
+    background-color: rgba(255, 255, 255, 0.6);
+    backdrop-filter: blur(5px);
+    padding: 10%;
+    height: 100%;
+    opacity: 0%;
+    font-size: 0.1em;
+    transition: all 0.5s ease-in-out;
+}
+
+.IUTTitle{
+    color: var(--quinary-color);
+    text-align: center;
+}
