+++ src/app/iut/iut.component.html
@@ -1,0 +1,28 @@
+<div id="IUT">
+    <div class="subTitle">
+      <h2>IUT2 Grenoble</h2>
+      <h3>Year 2020-2022 <br> DUT Informatique - 11e de promotion</h3>
+    </div>
+    <!-- Add the sliders -->
+    <div class="IUTSliderContainer">
+      <div class="IUTSlider Classproject">
+        <div class="IUTText">
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut in lectus et justo consequat rhoncus ut et orci. Sed pellentesque nunc non urna interdum tristique. Pellentesque a augue ante. Pellentesque eleifend malesuada lorem, in posuere diam. Mauris dictum maximus dui, at accumsan lorem laoreet elementum. Suspendisse sollicitudin aliquam ex vel venenatis. In porttitor, arcu nec porttitor dapibus, neque sapien interdum diam, ornare sodales nulla risus non nisi. Aliquam vitae quam tincidunt, scelerisque augue sit amet, pulvinar nisi. Praesent euismod ante eget tincidunt lacinia.
+        </div>
+        <!-- <img src="../assets/images/class_project.png"> -->
+        <h3 class="IUTTitle">Class projects</h3>
+      </div>
+      <div class="IUTSlider Graduationproject">
+        <div class="IUTText">
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut in lectus et justo consequat rhoncus ut et orci. Sed pellentesque nunc non urna interdum tristique. Pellentesque a augue ante. Pellentesque eleifend malesuada lorem, in posuere diam. Mauris dictum maximus dui, at accumsan lorem laoreet elementum. Suspendisse sollicitudin aliquam ex vel venenatis. In porttitor, arcu nec porttitor dapibus, neque sapien interdum diam, ornare sodales nulla risus non nisi. Aliquam vitae quam tincidunt, scelerisque augue sit amet, pulvinar nisi. Praesent euismod ante eget tincidunt lacinia.
+        </div>
+        <h3 class="IUTTitle">Graduation project</h3>
+      </div>
+      <div class="IUTSlider Bdeproject">
+        <div class="IUTText">
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut in lectus et justo consequat rhoncus ut et orci. Sed pellentesque nunc non urna interdum tristique. Pellentesque a augue ante. Pellentesque eleifend malesuada lorem, in posuere diam. Mauris dictum maximus dui, at accumsan lorem laoreet elementum. Suspendisse sollicitudin aliquam ex vel venenatis. In porttitor, arcu nec porttitor dapibus, neque sapien interdum diam, ornare sodales nulla risus non nisi. Aliquam vitae quam tincidunt, scelerisque augue sit amet, pulvinar nisi. Praesent euismod ante eget tincidunt lacinia.
+        </div>
+        <h3 class="IUTTitle">BDE project</h3>
+      </div>
+    </div>
+  </div>
