+++ src/app/iut/iut.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-iut',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './iut.component.html',
+  styleUrl: './iut.component.css'
+})
+export class IutComponent {
+
+}
