+++ src/app/projects/projects.component.html
@@ -1,0 +1,62 @@
+<div id="Projects">
+    <!-- <div id="ProjectsTopWhiteBlur"></div> -->
+  
+    <!-- <img id="ProjectsImage" src="assets/images/PersonnalProjects-low.png" alt="Countact Image"> -->
+  
+    <div id="ProjectsBody">
+      <div class="subTitle">
+        <h2>Personal projects</h2>
+        <h3>Starting during my first years at university</h3>
+        <div class="subTitletext">
+          <p>Lorem ipsum dolor sit amet, consectetur</p>
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum lacinia efficitur condimentum. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum lacinia efficitur condimentum. </p>
+        </div>
+      </div>
+      
+      <!-- Add the sliders -->
+      <div class="ProjectsSlider">
+  
+        <!-- Notion -->
+        <div class="ProjectsSliderItem">
+          <div class="SliderItemImage">
+            <img src="../assets/images/Notion_Automation.png">
+            <div class="SliderImageTitle">
+              2023
+            </div>
+          </div>
+          <div class="SliderItemTitle">
+            Notion Automation
+          </div>
+        </div>
+  
+        <!-- Incongrue App -->
+        <div class="ProjectsSliderItem">
+          <div class="SliderItemImage">
+            <img src="../assets/images/Incongrue_App.png">
+            <div class="SliderImageTitle">
+              2023
+            </div>
+          </div>
+          <div class="SliderItemTitle">
+            Incongrue App
+          </div>
+        </div>
+  
+        <!-- TripPics -->
+        <div class="ProjectsSliderItem">
+          <div class="SliderItemImage">
+            <img src="../assets/images/TripPics.png">
+            <div class="SliderImageTitle">
+              2022
+            </div>
+          </div>
+          <div class="SliderItemTitle">
+            TripPics
+          </div>
+        </div>
+  
+      </div>
+    </div>
+    
+    <!-- <div id="ProjectsBottomWhiteBlur"></div> -->
+  </div>
