+++ src/app/resume/resume.component.css
@@ -1,0 +1,39 @@
+#Resume{
+    width: 70%;
+    background-color: var(--quinary-color);
+    border-radius: 15px;
+    transform: translateY(-25px);
+    margin: auto;
+    display: flex;
+    flex-direction: row;
+    filter: drop-shadow(0px 4px 4px rgba(0,0,0,0.5));
+    cursor: pointer;
+    position: relative;
+    z-index: 3;
+}
+
+#ResumeText{
+    width: 30%;
+    max-height: 100%;
+    padding: 2% 4%;
+    font-family: 'Roboto', sans-serif;
+}
+
+#ResumeImage{
+    width: 70%;
+    max-height: 30vh;
+    padding: 2%;
+    transition: all 1s ease-in-out;
+}
+
+#ResumeImage.expand, #Resume.expand{
+    max-height: 200vh;
+}
+
+#ResumeImage img{
+    width: 100%;
+    max-height: 100%;
+    object-fit: cover;
+    object-position: top;
+    border-radius: 10px;
+}
