+++ src/app/resume/resume.component.html
@@ -1,0 +1,11 @@
+<div id="Resumebackground">
+    <div id="Resume" >
+      <div id="ResumeText">
+        <h3>Here is my resume</h3>
+        <p>If you want your visit to be quick and know already why you should choose me</p>
+      </div>
+      <div id="ResumeImage">
+        <a data-fancybox id="ResumeLightbox" class="iframe" href="../assets/pdf/Resume Symeon.pdf"><img id="ResumeClickable" src="assets/images/ResumeImage.png" alt="Resume Image"></a>
+      </div>
+    </div>
+  </div>
