+++ src/app/scu/scu.component.css
@@ -1,0 +1,19 @@
+#SCU{
+    width: 100%;
+    height: 70vh;
+    display: flex;
+    margin: auto;
+    overflow-y: clip;
+    overflow-x: clip;
+}
+
+#SCUBody{
+    padding: 10%;
+}
+
+#SCUImage{
+    width: 50%;
+    height: 50%;
+    align-self: end;
+    padding: 10% 0;
+}
