+++ src/app/scu/scu.component.html
@@ -1,0 +1,13 @@
+<div id="SCU">
+    <div id="SCUBody">
+      <div class="subTitle">
+        <h2>Southern Cross university</h2>
+        <h3>Year 2023-2024 <br> Bachelor in information technology</h3>
+      </div>
+      <div>
+        <p>Project A</p>
+        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum lacinia efficitur condimentum. Duis aliquet rhoncus diam, eu tempor lacus ullamcorper at. Duis at luctus dui. Praesent ultrices blandit ante, ut dapibus sem aliquet a. Nam pharetra turpis velit, et bibendum sapien pellentesque et. Phasellus odio diam, consectetur quis tincidunt vel, porta nec arcu.</p>
+      </div>
+    </div>
+    <img id="SCUImage" src="assets/images/SCU.png" alt="SCU Image">
+</div>
