+++ src/app/scu/scu.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-scu',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './scu.component.html',
+  styleUrl: './scu.component.css'
+})
+export class ScuComponent {
+
+}
