+++ src/app/transition/transition.component.css
@@ -1,0 +1,57 @@
+#Transition{
+    height: 70vh;
+    width: 100%;
+    display: flex;
+    align-items: center;
+}
+
+#TransitionText{
+    width: 30%;
+    margin-left: 10%;
+}
+
+#TransitionImages{
+    max-width: 60%;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+
+    /* Shadow */
+    filter: drop-shadow(0px 4px 4px rgba(0,0,0,0.5));
+
+    /* Parallax */
+    /* overflow-x: hidden;
+    overflow-y: auto;
+    perspective: 300px;
+    position: relative;
+    transform-style: preserve-3d; */
+}
+
+#TransitionImages img{
+    max-width: 40%;
+    max-height: 40%;
+    border-radius: 20px;
+}
+
+#TransitionImages > #TI1{
+    transform: translateX(40%) translateY(-60%);
+    z-index: 2;
+    width: 314px;
+    /* Parallax */
+}
+
+#TransitionImages > #TI2{
+    z-index: 1;
+    width: 314px;
+    /* Parallax */
+}
+
+#TransitionImages > #TI3{
+    max-width: 30%;
+    max-height: 30%;
+    transform: translateX(-40%) translateY(50%);
+    z-index: 2;
+    width: 235px;
+    /* Parallax */
+}
