+++ src/app/transition/transition.component.html
@@ -1,0 +1,14 @@
+<div id="Transition">
+    <div id="TransitionText">
+      <h3>If you want to know more about me</h3>
+      <p>Here is my work</p>
+    </div>
+    <div id="TransitionImages">
+      <img id="TI1" class="top_layer layer" src="../assets/images/surfers.jpg">
+      <img id="TI2" class="bottom_layer layer" src="../assets/images/tower.jpg">
+      <img id="TI3" class="mid_layer layer" src="../assets/images/bridge.jpg">
+    </div>
+  </div>
+  <div id="AboutMe">
+  
+  </div>
