+++ src/app/workwithme/workwithme.component.html
@@ -1,0 +1,40 @@
+<div id="WorkWithMeBackground">
+    <div id="WorkWithMe">
+      <div class="subTitle">
+        <h2>Work with me</h2>
+      </div>
+      <div id="WWMText">
+        <div id="WhatICanDo">
+          <div>
+            <h3>What I can do</h3>
+            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum lacinia efficitur condimentum. Duis aliquet rhoncus diam, eu tempor lacus ullamcorper at. Duis at luctus dui. Praesent ultrices blandit ante, ut dapibus sem aliquet a. Nam pharetra turpis velit, et bibendum sapien pellentesque et. Phasellus odio diam, consectetur quis tincidunt vel, porta nec arcu.</p>
+          </div>
+          <div id="WhatICanDoTechno">
+            <img src="../assets/images/logo_html.png">
+            <img src="../assets/images/logo_css.png">
+            <img src="../assets/images/logo_javascript.png">
+            <img src="../assets/images/logo_angular.png">
+            <img src="../assets/images/logo_php.png">
+            <img src="../assets/images/logo_react.png">
+            <img src="../assets/images/logo_c.png">
+            <img src="../assets/images/logo_django.png">
+            <img src="../assets/images/logo_python.png">
+            <img src="../assets/images/logo_postgresql.png">
+            <img src="../assets/images/logo_docker.png">
+            <img src="../assets/images/logo_aws.png">
+            <img src="../assets/images/logo_notion.png">
+            <img src="../assets/images/logo_kotlin.png">
+            <img src="../assets/images/logo_n8n.png">
+            <img src="../assets/images/logo_java.png">
+          </div>
+        </div>
+        <div id="ContactMe">
+          <h3>Contact me</h3>
+          <input type="text" placeholder="Name">
+          <input type="text" placeholder="Email">
+          <textarea placeholder="Message" class="FreeText"></textarea>
+          <button>Send</button>
+        </div>
+      </div>
+    </div>
+  </div>
