+++ src/app/workwithme/workwithme.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-workwithme',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './workwithme.component.html',
+  styleUrl: './workwithme.component.css'
+})
+export class WorkwithmeComponent {
+
+}
